Fixup the dijkstra's algorithm so it actually works properly
Moved finding the next node into it's own function, as it got quite
large on it's own.
This code is looking might messy now though. It might need a bunch of
refactoring to remove some edge cases that wont ever happen.

I shall leave the debugish prints in there, because they're a
nice way of showing whats going on.